chore: cleanup a3c agent API
