+++ libraries/mathy_python/mathy/agents/__init__.py
@@ -1,3 +1,6 @@
 from ..util import assert_tensorflow_installed
 
 assert_tensorflow_installed()
+
+from .agent import A3CAgent # noqa
+from .config import AgentConfig # noqa
