--- libraries/mathy_python/mathy/agents/a3c/__init__.py
@@ -1,4 +1,0 @@
-from .config import *  # noqa
-from .agent import *  # noqa
-from .worker import *  # noqa
-
