+++ libraries/mathy_python/mathy/agents/agent.py
@@ -6,20 +6,20 @@ from typing import List
 import gym
 import numpy as np
 
-from ...teacher import Teacher
-from ..policy_value_model import get_or_create_policy_model, PolicyValueModel
-from .config import A3CConfig
+from ..teacher import Teacher
+from .policy_value_model import get_or_create_policy_model, PolicyValueModel
+from .config import AgentConfig
 from .worker import A3CWorker
-from ...envs.gym import MathyGymEnv
-from ...state import observations_to_window
+from ..envs.gym import MathyGymEnv
+from ..state import observations_to_window
 
 
 class A3CAgent:
 
-    args: A3CConfig
+    args: AgentConfig
     global_model: PolicyValueModel
 
-    def __init__(self, args: A3CConfig, env_extra: dict = None):
+    def __init__(self, args: AgentConfig, env_extra: dict = None):
         import tensorflow as tf
 
         self.args = args
