+++ libraries/mathy_python/tests/test_api.py
@@ -1,7 +1,7 @@
 import pytest
 import tensorflow as tf
 
-from mathy.agents.base_config import BaseConfig
+from mathy.agents.config import AgentConfig
 from mathy.agents.policy_value_model import PolicyValueModel
 from mathy.api import Mathy, MathyAPISwarmState
 
@@ -11,11 +11,11 @@ def test_mathy_policy_value_subclass_error():
     # NOTE: if this test is breaking your model, let's talk about it on Github
     model = tf.keras.Model()
     with pytest.raises(ValueError):
-        mt = Mathy(model=model, config=BaseConfig())
+        mt = Mathy(model=model, config=AgentConfig())
 
 
 def test_mathy_with_model_and_config():
-    config = BaseConfig()
+    config = AgentConfig()
     model = PolicyValueModel(config, predictions=2)
     mt = Mathy(model=model, config=config)
 
@@ -30,11 +30,11 @@ def test_api_mathy_constructor():
 
     # Model must be PVM
     with pytest.raises(ValueError):
-        Mathy(model=dict(), config=BaseConfig())  # type:ignore
-    # Model Config must extend BaseConfig
+        Mathy(model=dict(), config=AgentConfig())  # type:ignore
+    # Model Config must extend AgentConfig
     with pytest.raises(ValueError):
         Mathy(model=PolicyValueModel(), config=dict())  # type:ignore
 
     # Config by itself must be SwarmConfig()
     with pytest.raises(ValueError):
-        Mathy(config=BaseConfig())
+        Mathy(config=AgentConfig())
