+++ libraries/website/docs/snippets/ml/a3c_profiling.py
@@ -1,14 +1,14 @@
 #!pip install gym
 import os
 from mathy.cli import setup_tf_env
-from mathy.agents.a3c import A3CAgent, A3CConfig
+from mathy.agents.a3c import A3CAgent, AgentConfig
 import shutil
 import tempfile
 
 model_folder = tempfile.mkdtemp()
 setup_tf_env()
 
-args = A3CConfig(
+args = AgentConfig(
     profile=True,
     max_eps=2,
     verbose=True,
