+++ libraries/website/docs/snippets/ml/a3c_training_mcts_worker_n.py
@@ -1,13 +1,13 @@
 #!pip install gym
 from mathy.cli import setup_tf_env
-from mathy.agents.a3c import A3CAgent, A3CConfig
+from mathy.agents.a3c import A3CAgent, AgentConfig
 import shutil
 import tempfile
 
 model_folder = tempfile.mkdtemp()
 setup_tf_env()
 
-args = A3CConfig(
+args = AgentConfig(
     action_strategy="mcts_worker_n",
     max_eps=1,
     verbose=True,
