+++ libraries/website/docs/snippets/ml/policy_value_basic.py
@@ -1,12 +1,12 @@
 import tensorflow as tf
 
 from mathy import envs
-from mathy.agents.base_config import BaseConfig
+from mathy.agents.config import AgentConfig
 from mathy.agents.policy_value_model import PolicyValueModel
 from mathy.env import MathyEnv
 from mathy.state import MathyObservation, observations_to_window
 
-args = BaseConfig()
+args = AgentConfig()
 env: MathyEnv = envs.PolySimplify()
 observation: MathyObservation = env.state_to_observation(env.get_initial_state()[0])
 model = PolicyValueModel(args, predictions=env.action_size)
