+++ libraries/website/docs/snippets/ml/policy_value_env_features.py
@@ -1,10 +1,10 @@
 from mathy import envs
-from mathy.agents.base_config import BaseConfig
+from mathy.agents.config import AgentConfig
 from mathy.agents.policy_value_model import PolicyValueModel
 from mathy.env import MathyEnv
 from mathy.state import MathyObservation, observations_to_window
 
-args = BaseConfig(use_env_features=True)
+args = AgentConfig(use_env_features=True)
 env: MathyEnv = envs.PolySimplify()
 observation: MathyObservation = env.state_to_observation(env.get_initial_state()[0])
 model = PolicyValueModel(args, predictions=env.action_size)
